requirements.txt
# Core App
streamlit
pandas
plotly
openpyxl
xlsxwriter
pillow
# Fuzzy Matching
rapidfuzz

# Google & OpenAI API
google-api-python-client
google-auth
google-generativeai
openai

# Utilities
pyshorteners
